Implemented the mqttBridge with simple command usage
The script to connect to IoT Core through mqtt has been implemented.
It is able to respond to a couple of basic commands that might be
needed. In the future more commands can be implemented, as well as
logic to deal with new configurations which can also be received
through the mqtt connection.

For now, a simulated payload is published.