Made changes to simplify mqtt bridge
Sometimes I want to run the loop without publishing messages, so a
quick way to turn off publishing was implemented instead of commenting
out the publish command. This way, it being disabled is not confused
with a connection or similar problem.

The payload was modified slightly, all the measurements are now
attached to the "measurements" keyword, packed in a dictionary.
The timestamp is added as a string as a different keyword. I was
previously adding the deviceId as a key, but that is no longer
necessary as they come as custom attributes on the pub/sub
message.